Se quita logica de socket.io de app.js
