+++ .env.model
@@ -1,4 +1,7 @@
 PORT=
 
-DB_MONGO_URL=
+DB_MONGO_USER=
+DB_MONGO_PASS=
+DB_MONGO_CLUSTER=
+DB_MONGO_APP=
 DB_MONDO_DB=
