+++ package.json
@@ -16,6 +16,7 @@
     "express": "^4.18.2",
     "express-handlebars": "^7.1.2",
     "mongoose": "^8.2.0",
+    "morgan": "^1.10.0",
     "nodemon": "^3.0.3",
     "socket.io": "^4.7.4"
   },
