+++ src/routes/views/cart.view.js
@@ -1,4 +1,3 @@
-//import CartManager from "../../dao/fileSystem/cartManager.js";
 import CartManager from "../../dao/mongoDB/cartManager.js"
 import { Router } from "express";
 const cartManager = new CartManager();
