+++ src/routes/views/chat.view.js
@@ -1,5 +1,4 @@
 import { Router } from "express";
-//import ChatManager from "../../dao/fileSystem/chatManager.js";
 import ChatManager from "../../dao/mongoDB/chatManager.js";
 
 const router = Router();
