+++ src/views/partials/formDelete.handlebars
@@ -1,7 +1,7 @@
 <div style="max-width: 100%">
     <form id="realTimeFormDelete">
         <label class="form-label">ID</label>
-        <input type="number" name="id" class="form-control" id="">
+        <input type="text" name="id" class="form-control" id="">
         <br>
         <button type="submit" class="btn btn-primary mb-4">Delete</button>
     </form>
